p4i integration Integrate 275663 from 4.3 to main: Fixes:    Make this compile on OS X. Details:  sizeof is defined as a ulong (at least on the mac) so, cast them both to the same type.
[git-p4: depot-paths = "//depot/qt/main/": change = 275664]

+++ tests/auto/qdom/tst_qdom.cpp
@@ -1648,7 +1648,7 @@ static const QChar umlautName[] =
  */
 void tst_QDom::germanUmlautToByteArray() const
 {
-    QCOMPARE(sizeof(umlautName) /  sizeof(QChar), unsigned(3));
+    QCOMPARE(ulong(sizeof(umlautName) /  sizeof(QChar)), ulong(3));
     const QString name(umlautName, 3);
 
     QDomDocument d;
